Add a base JNI class for WPIMath.
